% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smrt_ccs.R
\name{smrt_ccs}
\alias{smrt_ccs}
\title{Get CCS data from a particular report}
\usage{
smrt_ccs(baseurl, token, datasetuid)
}
\arguments{
\item{baseurl}{URL of the SMRTLink installation, e.g. https://servername:8243}

\item{token}{A token obtained with \code{smrt_token()}}

\item{datasetuid}{Dataset UUID}
}
\value{
A tibble with report information
}
\description{
The SMRTLink API endpoint used is \verb{/smrt-link/datasets/\{datasetType\}/\{datasetUUID\}/reports}
Dataset types can be obtained a GET request at \verb{/smrt-link/dataset-types}, the most relevant are
\code{subreads}, \code{ccsreads} and \code{transcripts}.
The access token can be obtained with \code{smrt_token()}
}
