% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smrt_gettoken.R
\name{smrt_gettoken}
\alias{smrt_gettoken}
\title{Obtain access token from a PacBio SMRTLink server v12.0}
\usage{
smrt_gettoken(baseurl, user, pass)
}
\arguments{
\item{baseurl}{URL of the SMRTLink installation, e.g. https://servername:8243}

\item{user}{Username}

\item{pass}{Password}
}
\value{
A list, the access token is under \code{access_token}
}
\description{
Obtain access token from a PacBio SMRTLink server v12.0
}
