% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smrt_runs.R
\name{smrt_runs}
\alias{smrt_runs}
\title{Get run information from a PacBio SMRTLink server v12.0}
\usage{
smrt_runs(baseurl, user, pass)
}
\arguments{
\item{baseurl}{URL of the SMRTLink installation, e.g. https://servername:8243}

\item{user}{Username}

\item{pass}{Password}
}
\value{
A tibble with run information
}
\description{
The API endpoint used is \verb{SMRTLink/1.0.0/smrt-link/runs}. This function uses \code{smrt_token()} internally to obtain
an access token, meaning that a new token is obtained with each invocation.
}
