% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smrt_subreads.R
\name{smrt_subreads}
\alias{smrt_subreads}
\title{Get subreads data for a particular cell}
\usage{
smrt_subreads(baseurl, token, uid)
}
\arguments{
\item{baseurl}{URL of the SMRTLink installation, e.g. https://servername:8243}

\item{token}{A token obtained with \code{smrt_token()}}

\item{celluid}{cell unique id}
}
\value{
A tibble with report information
}
\description{
The SMRTLink API endpoint used is \verb{/smrt-link/datasets/\{datasetType\}/\{datasetUUID\}/reports}
Dataset types can be obtained a GET request at \verb{/smrt-link/dataset-types}, the most relevant are
\code{subreads}, \code{ccsreads} and \code{transcripts}.
The access token can be obtained with \code{smrt_token()}
}
