% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smrt_token.R
\name{smrt_token}
\alias{smrt_token}
\title{Obtain access token from a PacBio SMRTLink server v12.0}
\usage{
smrt_token(baseurl, user, pass)
}
\arguments{
\item{baseurl}{URL of the SMRTLink installation, e.g. https://servername:8243}

\item{user}{Username}

\item{pass}{Password}
}
\value{
A list, the access token is under \code{access_token}
}
\description{
The API endpoint used is \verb{SMRTLink/1.0.0/token}
}
